Major changes: .inex, existing file validation, implelemt open file and remove file using filename from user

+++ src/loadDatabase.c
@@ -1,5 +1,6 @@
 #include<stdio.h>
 #include<stdlib.h>
+#include<string.h>
 
 #include"headers/definition.h"
 #include"headers/customFunctions.h"
@@ -41,22 +42,41 @@ Database* createDatabase(int *returnCode)
 Database* openDatabase(int *returnCode) 
 {
     Database *db = NULL;
-    char file_name[64] = "test.bin";
+    char file_name[64];
 
     if (returnCode == NULL) {
         printErrorMessage(-2);
         return db;
     }
 
-    printf("\n\tMESSAGE: <UNDER DEVELOPMENT>, Trying to open 'test.bin'\n");
+    fprintf(stdout,"\n\tPlease Enter File name to OPEN: ");
 
-    *returnCode = 0;
+    *returnCode = getStringInput(stdin,file_name,64);
+    if (*returnCode <= 0) {
+        if (*returnCode == -1) 
+            *returnCode = -3;
+        if (*returnCode == 0)
+            *returnCode = -4;
+        printErrorMessage(*returnCode);
+        return NULL;
+    }
+
+    if (*returnCode >= 50) {
+        /*---ERROR MESSAGE---*/
+        fprintf(stdout,"\n\tERROR: Filename should not be more than 50 characters\n");
+        *returnCode = -1;
+        return NULL;
+    }
+
+    /*---UNSAFE FUNCTION---*/
+    strcat(file_name,".inex");
 
     db = openFile(file_name);
     
     if (db == NULL) {
         *returnCode = -1;
     } else {
+        *returnCode = 0;
         db->dbMetaData.metaData.isSaved = 1;
         db->dbMetaData.metaData.isModified = 0;
     }
@@ -69,7 +89,27 @@ int deleteDatabase()
     char file_name[64] = "test.bin";
     int returnCode = 0;
 
-    printf("\n\tMESSAGE: <UNDER DEVELOPMENT>, Trying to delete 'test.bin'\n");
+    fprintf(stdout,"\n\tPlease Enter File name to DELETE: ");
+
+    returnCode = getStringInput(stdin,file_name,64);
+    if (returnCode <= 0) {
+        if (returnCode == -1) 
+            returnCode = -3;
+        if (returnCode == 0)
+            returnCode = -4;
+        printErrorMessage(returnCode);
+        return returnCode;
+    }
+
+    if (returnCode > 50) {
+        /*---ERROR MESSAGE---*/
+        fprintf(stdout,"\n\tERROR: Filename should not be more than 50 characters\n");
+        returnCode = -1;
+        return returnCode;
+    }
+
+    /*---UNSAFE FUNCTION---*/
+    strcat(file_name,".inex");
 
     if (returnCode = removeFile(file_name) != 0) 
         printErrorMessage(-10);
